feature/card change to different format
