+++ src/components/Card/Card.tsx
@@ -1,11 +1,10 @@
-import React, { ReactElement } from 'react'
-import CardBase, { CardBaseProps } from './components/CardBase'
-import CardFigure, { CardFigureProps } from './components/CardFigure'
-import CardNews, { CardNewsProps } from './components/CardNews'
-import CardVideo, { CardVideoProps } from './components/CardVideo'
+import React from 'react'
 
-export interface CardProps {
-    type: 'base' | 'news' | 'video' | 'figure'
+export interface CardBaseProps {
+    link: string
+    header: string
+    description?: string
+    image: CardImageProps
 }
 
 export interface CardImageProps {
@@ -13,22 +12,30 @@ export interface CardImageProps {
     alt?: string
 }
 
-export interface TypeProps {
-    [index: string]: ReactElement
-}
-
-const Card: React.FC<
-    CardProps & CardBaseProps & CardNewsProps & CardFigureProps & CardVideoProps
-> = (props): JSX.Element => {
-    const { type = 'base' } = props
-
-    const cardTypes: TypeProps = {
-        base: <CardBase {...props} />,
-        news: <CardNews {...props} />,
-        video: <CardVideo {...props} />,
-        figure: <CardFigure {...props} />,
-    }
-    return cardTypes[type || 'base']
-}
-
+const Card: React.FC<CardBaseProps> = ({
+    link,
+    header,
+    description,
+    image,
+}): JSX.Element => (
+    <article className='c-card' itemScope itemType='http://schema.org/Article'>
+        <a href={link} itemProp='url'>
+            <figure itemScope itemType='http://schema.org/ImageObject'>
+                <img src={image.src} alt={image.alt} loading='lazy' />
+            </figure>
+            <header>
+                <h3
+                    itemProp='name'
+                    dangerouslySetInnerHTML={{ __html: header }}
+                />
+                {description && (
+                    <p
+                        itemProp='description'
+                        dangerouslySetInnerHTML={{ __html: description }}
+                    />
+                )}
+            </header>
+        </a>
+    </article>
+)
 export default Card
