--- src/components/Card/components/CardBase.tsx
@@ -1,37 +1,0 @@
-import React from 'react'
-import { CardImageProps } from '../Card'
-
-export interface CardBaseProps {
-    link: string
-    header: string
-    description?: string
-    image: CardImageProps
-}
-
-const CardBase: React.FC<CardBaseProps> = ({
-    link,
-    header,
-    description,
-    image,
-}): JSX.Element => (
-    <article className='c-card' itemScope itemType='http://schema.org/Article'>
-        <a href={link} itemProp='url'>
-            <figure itemScope itemType='http://schema.org/ImageObject'>
-                <img src={image.src} alt={image.alt} loading='lazy' />
-            </figure>
-            <header>
-                <h3
-                    itemProp='name'
-                    dangerouslySetInnerHTML={{ __html: header }}
-                />
-                {description && (
-                    <p
-                        itemProp='description'
-                        dangerouslySetInnerHTML={{ __html: description }}
-                    />
-                )}
-            </header>
-        </a>
-    </article>
-)
-export default CardBase
